Delete model.go
